# remove SMT solver from RewriteStepState, use solver from arg. to evaluate