Supplies the SMT solver from the `performRewrite` instead of a (wrong) value from `RewriteStepState`.
We don't need the SMT solver in the `RewriteStepState` as it won't be modified, and the old code was also setting it to `Nothing`.