@jberthold 2024-07-29
| Test                                                  | HOTFIX-actually-use-smt-solver-for-equations time | master-613cb4984 time | (HOTFIX-actually-use-smt-solver-for-equations/master-613cb4984) time 
|-------------------------------------------------------|---------------------------------------------------|-----------------------|----------------------------------------------------------------------
| benchmarks/ecrecoverloop02-sigs-valid-spec.k          | 85.65                                             | 190.76                | 0.4489934996854687                                                   
| benchmarks/ecrecoverloop02-sig1-invalid-spec.k        | 85.63                                             | 187.36                | 0.4570345858240819                                                   
| benchmarks/ecrecoverloop02-sig0-invalid-spec.k        | 66.5                                              | 112.22                | 0.5925859918018178                                                   
| mcd/end-cash-pass-rough-spec.k                        | 154.24                                            | 224.16                | 0.6880799428979301                                                   
| erc20/ds/transferFrom-success-1-spec.k                | 53.48                                             | 64.6                  | 0.8278637770897833                                                   
| erc20/ds/transferFrom-success-2-spec.k                | 50.62                                             | 59.89                 | 0.8452162297545499                                                   
| erc20/ds/transfer-success-1-spec.k                    | 46.18                                             | 53.6                  | 0.8615671641791044                                                   
| erc20/ds/transfer-success-2-spec.k                    | 46.02                                             | 52.8                  | 0.8715909090909092                                                   
| mcd/flapper-yank-pass-rough-spec.k                    | 136.89                                            | 154.85                | 0.8840167904423635                                                   
| mcd/flopper-cage-pass-spec.k                          | 40.1                                              | 44.9                  | 0.8930957683741648                                                   
| mcd/end-pack-pass-rough-spec.k                        | 115.25                                            | 128.31                | 0.8982152599173876                                                   
| erc20/hkg/transferFrom-success-1-spec.k               | 47.87                                             | 51.66                 | 0.9266356949283778                                                   
| erc20/ds/transfer-failure-1-b-spec.k                  | 68.45                                             | 72.67                 | 0.9419292692995734                                                   
| erc20/hkg/transfer-success-1-spec.k                   | 42.27                                             | 44.86                 | 0.9422648238965672                                                   
| benchmarks/ecrecoverloop00-sig1-invalid-spec.k        | 77.43                                             | 81.98                 | 0.9444986582093193                                                   
| mcd/cat-file-addr-pass-rough-spec.k                   | 46.02                                             | 48.69                 | 0.9451632778804684                                                   
| erc20/hkg/transferFrom-success-2-spec.k               | 45.7                                              | 48.3                  | 0.9461697722567289                                                   
| erc20/ds/approve-success-spec.k                       | 42.61                                             | 45.03                 | 0.946258050188763                                                    
| erc20/hkg/approve-spec.k                              | 38.53                                             | 40.56                 | 0.9499506903353057                                                   
| erc20/ds/transferFrom-failure-1-b-spec.k              | 105.33                                            | 110.63                | 0.952092560788213                                                    
| erc20/hkg/transfer-success-2-spec.k                   | 41.38                                             | 43.21                 | 0.9576486924323073                                                   
| erc20/hkg/allowance-spec.k                            | 38.55                                             | 40.06                 | 0.9623065401897153                                                   
| mcd/flopper-dent-guy-same-pass-rough-spec.k           | 526.28                                            | 546.26                | 0.9634240105444294                                                   
| erc20/ds/transfer-failure-2-a-spec.k                  | 53.71                                             | 55.59                 | 0.966180967799964                                                    
| kontrol/test-allowchangestest-testallow_fail-0-spec.k | 41.8                                              | 40.38                 | 1.035165923724616                                                    
| benchmarks/ecrecoverloop00-sigs-valid-spec.k          | 55.24                                             | 53.23                 | 1.0377606612812325                                                   
| mcd/end-subuu-pass-spec.k                             | 48.03                                             | 41.54                 | 1.1562349542609534                                                   
| TOTAL                                                 | 2199.76                                           | 2638.1                | 0.8338425381903644  

:open_mouth: